refactor: create login service and move some logic to it
